Added new functions and a button listener
This commit adds a few comments to the JS file for better understanding and study purposes.
Adds a new conditional to the add function
Adds a new function that creates a list of buttons based on the pokemon objects. Created a event listener to the button so that it logs back the pokemon name to the console.